2025-04-28

+++ umh-core/pkg/communicator/models/mgmtconfig/kafkabroker.go
@@ -16,5 +16,5 @@ package mgmtconfig
 
 type KafkaBroker struct {
 	Ip   string `yaml:"ip"`
-	Port int    `yaml:"port"`
+	Port uint16 `yaml:"port"`
 }
